requirements.txt
anyio==2.0.2
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
black==20.8b1
bleach==3.3.0
build==0.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
configparser==5.0.1
contextvars==2.4
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
idna==2.10
immutables==0.14
importlib-metadata==3.4.0
iniconfig==1.1.1
ipykernel==5.4.3
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.17.2
Jinja2==2.11.3
joblib==1.0.0
json5==0.9.5
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-server==1.3.0
jupyterlab==3.0.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.4.1
numpy==1.18.2
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
pathspec==0.8.1
pep517==0.9.1
pexpect==4.8.0
pickleshare==0.7.5
pip-autoremove==0.9.1
pluggy==0.13.1
progressbar2==3.53.1
prometheus-client==0.9.0
prompt-toolkit==3.0.14
ptyprocess==0.7.0
py==1.10.0
pycparser==2.20
pygam==0.8.0
Pygments==2.7.4
pylint==2.6.0
pymongo==3.11.3
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-githooks==1.0.6
python-utils==2.5.6
pytz==2021.1
pyzmq==22.0.2
regex==2020.11.13
requests==2.25.1
scikit-learn==0.24.1
scipy==1.5.1
Send2Trash==1.5.0
six==1.15.0
sklearn-pandas==2.2.0
sniffio==1.2.0
Sphinx==4.1.2
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.60.0
traitlets==4.3.3
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
zipp==3.4.0
